Codechange: Give sprite encoders a hint which colour components of a sprite are filled with useful information.
